feature(core-api): add unit and integration tests for the launch module

+++ moonkeyeu-core-api/src/main/java/com/moonkeyeu/core/api/launch/dto/filters/FiltersDTO.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @JsonPropertyOrder({ "id", "name" })
 public class FiltersDTO implements DTOEntity {
     @JsonProperty("id")
-    private Integer filterId;
+    private Long filterId;
     @JsonProperty("name")
     private String filterName;
 }
